refactor: rename newMenuLink and newMenuSection to burger instead of new
